Restructured assets for convenience in packaging
