--- main.py
@@ -1,57 +1,0 @@
-"""
-##########################################################################
-*
-*   Copyright © 2019-2021 Fedora Websites and Apps Team
-*
-*   This program is free software: you can redistribute it and/or modify
-*   it under the terms of the GNU General Public License as published by
-*   the Free Software Foundation, either version 3 of the License, or
-*   (at your option) any later version.
-*
-*   This program is distributed in the hope that it will be useful,
-*   but WITHOUT ANY WARRANTY; without even the implied warranty of
-*   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-*   GNU General Public License for more details.
-*
-*   You should have received a copy of the GNU General Public License
-*   along with this program.  If not, see <https://www.gnu.org/licenses/>.
-*
-##########################################################################
-"""
-
-import click
-from flask import Flask, \
-    render_template
-
-main = Flask(__name__)
-
-
-@main.get("/")
-def mainpage():
-    return render_template("mainpage.html")
-
-
-@click.command()
-@click.option("-p", "--portdata", "portdata", help="Set the port value [0-65536]", default="9696")
-@click.option("-6", "--ipprotv6", "netprotc", flag_value="ipprotv6", help="Start the server on an IPv6 address")
-@click.option("-4", "--ipprotv4", "netprotc", flag_value="ipprotv4", help="Start the server on an IPv4 address")
-@click.version_option(version="0.1.0", prog_name="Fedora Easyfix")
-def mainfunc(portdata, netprotc):
-    """
-    A collection of self-contained and well-documented issues for newcomers to start contributing with
-    """
-    print(" * Starting Fragment...")
-    print(" * Port number : " + str(portdata))
-    netpdata = ""
-    if netprotc == "ipprotv6":
-        print(" * IP version  : 6")
-        netpdata = "::"
-    elif netprotc == "ipprotv4":
-        print(" * IP version  : 4")
-        netpdata = "0.0.0.0"
-    main.config["TEMPLATES_AUTO_RELOAD"] = True
-    main.run(port=portdata, host=netpdata)
-
-
-if __name__ == "__main__":
-    mainfunc()
