+++ src/fedora_easyfix/static/imgs/USAGE
